Refactoring from utils to mapper package
