+++ src/main/java/io/strimzi/kafka/bridge/mqtt/mapper/MqttKafkaMapper.java
@@ -2,9 +2,7 @@
  * Copyright Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.strimzi.kafka.bridge.mqtt.core;
-
-import io.strimzi.kafka.bridge.mqtt.utils.MappingRule;
+package io.strimzi.kafka.bridge.mqtt.mapper;
 
 import java.util.List;
 import java.util.ArrayList;
